third_party/WebKit/LayoutTests/http/tests/serviceworker/postmessage.html

<!DOCTYPE html>
<title>Service Worker: postMessage</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
async_test(function(t) {
    var scope = 'resources/blank.html';
    var registration;
    var worker;
    service_worker_unregister_and_register(
        t, 'resources/postmessage-worker.js', scope)
      .then(function(r) {
          registration = r;
          return wait_for_update(t, registration);
        })
      .then(function(sw) {
          worker = sw;
          var messageChannel = new MessageChannel();
          messageChannel.port1.onmessage = t.step_func(onMessage);
          sw.postMessage({port: messageChannel.port2}, [messageChannel.port2]);
          sw.postMessage({value: 1});
          sw.postMessage({value: 2});
          sw.postMessage({done: true});
        })
      .catch(unreached_rejection(t));

    var result = [];
    var expected = [
      'Acking value: 1',
      'Acking value: 2',
    ];

    function onMessage(e) {
      var message = e.data;
      if (message === 'quit') {
        assert_array_equals(result, expected,
                            'Worker should post back expected values.');
        var sawRedundant = new Promise(t.step_func(function(resolve) {
            worker.onstatechange = t.step_func(function() {
                assert_equals(worker.state, 'redundant');
                assert_throws(
                  {name:'InvalidStateError'},
                  function() { worker.postMessage(""); },
                  'Calling postMessage on a redundant ServiceWorker should ' +
                  'throw InvalidStateError.');
                resolve();
              });
          }));
        registration.unregister();
        sawRedundant.then(t.step_func(function() { t.done(); }));
      } else {
        result.push(message);
      }
    };
  }, 'postMessage to a ServiceWorker (and back via MessagePort)');
</script>
